Clean soccer.go & create sportmonks package
