+++ server/Gopkg.lock
@@ -13,9 +13,15 @@
   revision = "24fca303ac6da784b9e8269f724ddeb0b2eea5e7"
   version = "v1.5.0"
 
+[[projects]]
+  branch = "master"
+  name = "github.com/mitchellh/mapstructure"
+  packages = ["."]
+  revision = "d0303fe809921458f417bcf828397a65db30a7e4"
+
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "11130969cb0149a105d78f30213e9329cfb4ace667952be38da5f8a9c8afa0f9"
+  inputs-digest = "9e46d2335e153433d2885c9b4dc6e34db67c1aebc8b5976b95b3fa037e929306"
   solver-name = "gps-cdcl"
   solver-version = 1
