+++ server/sportmonks/fixtures.go
@@ -1,0 +1,1 @@
+package sportmonks
